refactor: create layout components
