+++ src/styles/layouts/components/Footer.ts
@@ -1,0 +1,16 @@
+import styled from 'styled-components';
+
+export const Container = styled.footer`
+  display: flex;
+  align-items: center;
+  justify-content: flex-end;
+
+  padding: 1rem 0;
+`;
+
+export const SwitchIconContainer = styled.div`
+  display: flex;
+  align-items: center;
+  justify-content: center;
+`;
+
